GetLap() Fix
lap variable was a float for some reason, is now an int as intended.

+++ Assets/Scripts/MotionController.cs
@@ -31,7 +31,7 @@ public class MotionController : MonoBehaviour {
 	private GameObject[] trackers; // Array of trackers on the track for respawn and checkpoint functions
 	private GameObject closestTracker; // the GameObject container for the tracker closest to the vehicle at all times.
 	private GameObject finishLine; // The collision object for crossing the finish line. 
-	private float lap = 1; // Laps, increments when passing 3 checkpoint and finishline objects.
+	private int lap = 1; // Laps, increments when passing 3 checkpoint and finishline objects.
 	private bool[] checks = new bool[]{false,false,false}; // Array of bools, each element turns true when checkpoints are reached.
